Вроде всё

+++ trpoMarkAnalizerProject/trpoMarkAnalizerProject/Form1.cs
@@ -31,6 +31,7 @@ namespace trpoMarkAnalizerProject
             ComboBoxAdd();
             ComboBoxAdd2();
             ComboBoxAdd3();
+            ComboBoxAdd4();
         }
 
         private void InitPage()
@@ -43,15 +44,23 @@ namespace trpoMarkAnalizerProject
             string query = "SELECT nameGroup, id FROM [Group]";
             OleDbDataAdapter da = new OleDbDataAdapter(query, _connection);
             DataTable ds = new DataTable();
-            da.Fill(ds);
-            comboBox3.DataSource = ds;
-            comboBox3.DisplayMember = "nameGroup";
-            comboBox3.ValueMember = "id";
-            comboBox3.SelectedIndex = -1;
+            da.Fill(ds);           
             comboBox1.DataSource = ds;
             comboBox1.DisplayMember = "nameGroup";
             comboBox1.ValueMember = "id";
             comboBox1.SelectedIndex = -1;
+        }
+
+        public void ComboBoxAdd4()
+        {
+            string query = "SELECT nameGroup, id FROM [Group]";
+            OleDbDataAdapter da = new OleDbDataAdapter(query, _connection);
+            DataTable ds = new DataTable();
+            da.Fill(ds);
+            groupStudentBox.DataSource = ds;
+            groupStudentBox.DisplayMember = "nameGroup";
+            groupStudentBox.ValueMember = "id";
+            groupStudentBox.SelectedIndex = -1;            
         }
 
         public void ComboBoxAdd2()
@@ -305,25 +314,46 @@ Data Source=Marks1.accdb;Persist Security Info=True");
 
         private void Save_button12_Click(object sender, EventArgs e)
         {
-            if (Add_label.Text == "Добавление")
+            try
             {
-                if ((lastName_textBox3.Text == "") || (firstName_textBox3.Text == "") || (sureName_textBox3.Text == "") || (aducation_comboBox2.Text == "") || (adress_textBox3.Text == "") || (phoneNumber.Text == ""))
+                if (Add_label.Text == "Добавление")
                 {
-                    MessageBox.Show("Заполните все поля");
-                }
-                else
-                {
-                    if ((Convert.ToDateTime(dateBirth.Text) < Convert.ToDateTime("01.01.1950")) || (Convert.ToDateTime(dateBirth.Text) > Convert.ToDateTime("01.01.1998")))
+                    if ((lastName_textBox3.Text == "") || (firstName_textBox3.Text == "") || (sureName_textBox3.Text == "") || (aducation_comboBox2.Text == "") || (adress_textBox3.Text == "") || (phoneNumber.Text == ""))
                     {
-                        MessageBox.Show("Укажите дату корректно");
+                        MessageBox.Show("Заполните все поля");
                     }
                     else
                     {
-                        int j = teacherGrid.RowCount - 2;
-                        int id = Convert.ToInt32(teacherGrid[0, j].Value.ToString());
-                        int newi = id + 1;
-                        string newid = newi.ToString();
-                        string query = $"INSERT INTO Teacher VALUES ('{newid}','{lastName_textBox3.Text}','{firstName_textBox3.Text}','{sureName_textBox3.Text}','{dateBirth.Text}', '{aducation_comboBox2.Text}', '{adress_textBox3.Text}', '{phoneNumber.Text}')";
+                        if ((Convert.ToDateTime(dateBirth.Text) < Convert.ToDateTime("01.01.1950")) || (Convert.ToDateTime(dateBirth.Text) > Convert.ToDateTime("01.01.1998")))
+                        {
+                            MessageBox.Show("Укажите дату корректно");
+                        }
+                        else
+                        {
+                            int j = teacherGrid.RowCount - 2;
+                            int id = Convert.ToInt32(teacherGrid[0, j].Value.ToString());
+                            int newi = id + 1;
+                            string newid = newi.ToString();
+                            string query = $"INSERT INTO Teacher VALUES ('{newid}','{lastName_textBox3.Text}','{firstName_textBox3.Text}','{sureName_textBox3.Text}','{dateBirth.Text}', '{aducation_comboBox2.Text}', '{adress_textBox3.Text}', '{phoneNumber.Text}')";
+                            OleDbCommand command = new OleDbCommand(query, _connection);
+                            command.ExecuteNonQuery();
+                            TeacherShow();
+                            panel1.Visible = false;
+                        }
+                    }
+                }
+                else
+                if (Add_label.Text == "Изменение")
+                {
+                    int index = teacherGrid.CurrentRow.Index;
+                    if (index == teacherGrid.RowCount - 1)
+                    {
+                        MessageBox.Show("Выберите заполненную сторку");
+                    }
+                    else
+                    {
+                        int id = Convert.ToInt32(teacherGrid[0, index].Value.ToString());
+                        string query = $"UPDATE Teacher SET lastName = '{lastName_textBox3.Text}', firstName = '{firstName_textBox3.Text}', sureName = '{sureName_textBox3.Text}', dateBirth = '{dateBirth.Text}', aducation = '{aducation_comboBox2.Text}', address = '{adress_textBox3.Text}', phoneNum = '{phoneNumber.Text}' WHERE Teacher.id = {id}";
                         OleDbCommand command = new OleDbCommand(query, _connection);
                         command.ExecuteNonQuery();
                         TeacherShow();
@@ -331,23 +361,9 @@ Data Source=Marks1.accdb;Persist Security Info=True");
                     }
                 }
             }
-            else
-            if (Add_label.Text == "Изменение")
+            catch (Exception a)
             {
-                int index = teacherGrid.CurrentRow.Index;
-                if (index == teacherGrid.RowCount - 1)
-                {
-                    MessageBox.Show("Выберите заполненную сторку");
-                }
-                else
-                {
-                    int id = Convert.ToInt32(teacherGrid[0, index].Value.ToString());
-                    string query = $"UPDATE Teacher SET lastName = '{lastName_textBox3.Text}', firstName = '{firstName_textBox3.Text}', sureName = '{sureName_textBox3.Text}', dateBirth = '{dateBirth.Text}', aducation = '{aducation_comboBox2.Text}', address = '{adress_textBox3.Text}', phoneNum = '{phoneNumber.Text}' WHERE Teacher.id = {id}";
-                    OleDbCommand command = new OleDbCommand(query, _connection);
-                    command.ExecuteNonQuery();
-                    TeacherShow();
-                    panel1.Visible = false;
-                }
+                MessageBox.Show(a.Message);
             }
         }
 
@@ -374,20 +390,24 @@ Data Source=Marks1.accdb;Persist Security Info=True");
 
         private void button7_Click(object sender, EventArgs e)
         {
-            panel1.Visible = false;
-            int index = teacherGrid.CurrentRow.Index;
-            if (index == teacherGrid.RowCount - 1)
+            var res = MessageBox.Show("Вы действительно хотите удалить запись?", "Удаление", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation);
+            if (res == DialogResult.Yes)
             {
-                MessageBox.Show("Выберите заполненную сторку");
+                panel1.Visible = false;
+                int index = teacherGrid.CurrentRow.Index;
+                if (index == teacherGrid.RowCount - 1)
+                {
+                    MessageBox.Show("Выберите заполненную сторку");
+                }
+                else
+                {
+                    int id = Convert.ToInt32(teacherGrid[0, index].Value.ToString());
+                    string query = $"DELETE FROM Teacher WHERE Teacher.id = {id}";
+                    OleDbCommand command = new OleDbCommand(query, _connection);
+                    command.ExecuteNonQuery();
+                }
+                TeacherShow();
             }
-            else
-            {
-                int id = Convert.ToInt32(teacherGrid[0, index].Value.ToString());
-                string query = $"DELETE FROM Teacher WHERE Teacher.id = {id}";
-                OleDbCommand command = new OleDbCommand(query, _connection);
-                command.ExecuteNonQuery();
-            }
-            TeacherShow();
         }
 
         private void textBox2_TextChanged(object sender, EventArgs e)
@@ -417,20 +437,24 @@ Data Source=Marks1.accdb;Persist Security Info=True");
 
         private void button4_Click(object sender, EventArgs e)
         {
-            panel3.Visible = false;
-            int index = subjectGrid.CurrentRow.Index;
-            if (index == subjectGrid.RowCount - 1)
+            var res = MessageBox.Show("Вы действительно хотите удалить запись?", "Удаление", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation);
+            if (res == DialogResult.Yes)
             {
-                MessageBox.Show("Выберите заполненную сторку");
+                panel3.Visible = false;
+                int index = subjectGrid.CurrentRow.Index;
+                if (index == subjectGrid.RowCount - 1)
+                {
+                    MessageBox.Show("Выберите заполненную сторку");
+                }
+                else
+                {
+                    int id = Convert.ToInt32(subjectGrid[0, index].Value.ToString());
+                    string query = $"DELETE FROM Subject WHERE Subject.id = {id}";
+                    OleDbCommand command = new OleDbCommand(query, _connection);
+                    command.ExecuteNonQuery();
+                }
+                SubjectShow();
             }
-            else
-            {
-                int id = Convert.ToInt32(subjectGrid[0, index].Value.ToString());
-                string query = $"DELETE FROM Subject WHERE Subject.id = {id}";
-                OleDbCommand command = new OleDbCommand(query, _connection);
-                command.ExecuteNonQuery();
-            }
-            SubjectShow();
         }
 
         private void button3_Click(object sender, EventArgs e)
@@ -459,44 +483,51 @@ Data Source=Marks1.accdb;Persist Security Info=True");
 
         private void button8_Click_1(object sender, EventArgs e)
         {
-            if (label11.Text == "Добавление")
+            try
             {
-                if ((comboBox4.Text == "") || (comboBox2.Text == ""))
+                if (label11.Text == "Добавление")
                 {
-                    MessageBox.Show("Заполните все поля");
+                    if ((comboBox4.Text == "") || (comboBox2.Text == ""))
+                    {
+                        MessageBox.Show("Заполните все поля");
+                    }
+                    else
+                    {
+                        int j = subjectGrid.RowCount - 2;
+                        int id = Convert.ToInt32(subjectGrid[0, j].Value.ToString());
+                        int newi = id + 1;
+                        string newid = newi.ToString();
+                        string query = $"INSERT INTO Subject(idTeacher, nameSub) VALUES ({comboBox4.SelectedValue},'{comboBox2.Text}')";
+                        OleDbCommand command = new OleDbCommand(query, _connection);
+                        command.ExecuteNonQuery();
+                        SubjectShow();
+                        panel3.Visible = false;
+                    }
                 }
                 else
+               if (label11.Text == "Изменение")
                 {
-                    int j = subjectGrid.RowCount - 2;
-                    int id = Convert.ToInt32(subjectGrid[0, j].Value.ToString());
-                    int newi = id + 1;
-                    string newid = newi.ToString();
-                    string query = $"INSERT INTO Subject(idTeacher, nameSub) VALUES ({comboBox4.SelectedValue},'{comboBox2.Text}')";
-                    OleDbCommand command = new OleDbCommand(query, _connection);
-                    command.ExecuteNonQuery();
-                    SubjectShow();
-                    panel3.Visible = false;
+                    int index = subjectGrid.CurrentRow.Index;
+                    if (index == subjectGrid.RowCount - 1)
+                    {
+                        MessageBox.Show("Выберите заполненную сторку");
+                    }
+                    else
+                    {
+                        int id = Convert.ToInt32(subjectGrid[0, index].Value.ToString());
+                        string query = $"UPDATE Subject SET idTeacher = '{comboBox4.SelectedValue}', nameSub = '{comboBox2.Text}' WHERE Subject.id = {id}";
+                        OleDbCommand command = new OleDbCommand(query, _connection);
+                        command.ExecuteNonQuery();
+                        SubjectShow();
+                        panel3.Visible = false;
+                    }
                 }
             }
-            else
-           if (label11.Text == "Изменение")
+            catch (Exception a)
             {
-                int index = subjectGrid.CurrentRow.Index;
-                if (index == subjectGrid.RowCount - 1)
-                {
-                    MessageBox.Show("Выберите заполненную сторку");
-                }
-                else
-                {
-                    int id = Convert.ToInt32(subjectGrid[0, index].Value.ToString());
-                    string query = $"UPDATE Subject SET idTeacher = '{comboBox4.SelectedValue}', nameSub = '{comboBox2.Text}' WHERE Subject.id = {id}";
-                    OleDbCommand command = new OleDbCommand(query, _connection);
-                    command.ExecuteNonQuery();
-                    SubjectShow();
-                    panel3.Visible = false;
-                }
+                MessageBox.Show(a.Message);
             }
-        }
+}
 
         private void comboBox4_SelectedIndexChanged(object sender, EventArgs e)
         {
@@ -525,5 +556,103 @@ Data Source=Marks1.accdb;Persist Security Info=True");
             textBox11.Text = "";
         }
 
+        private void button2_Click(object sender, EventArgs e)
+        {
+            label19.Text = "Добавление";
+            groupStudentBox.Text = "";
+            textBox9.Text = "";
+            textBox10.Text = "";
+            textBox8.Text = "";
+            textBox7.Text = "";
+            maskedTextBox4.Text = "";
+            panel4.Visible = true;
+        }
+
+        private void button11_Click(object sender, EventArgs e)
+        {
+            int index = studentGrid.CurrentRow.Index;
+            if (index == studentGrid.RowCount - 1)
+            {
+                MessageBox.Show("Выберите заполненную сторку");
+            }
+            else
+            {
+                label19.Text = "Изменение";
+                groupStudentBox.Text = studentGrid[1, index].Value.ToString();
+                textBox9.Text = studentGrid[2, index].Value.ToString();
+                textBox10.Text = studentGrid[3, index].Value.ToString();
+                textBox8.Text = studentGrid[4, index].Value.ToString();
+                textBox7.Text = studentGrid[5, index].Value.ToString();
+                maskedTextBox4.Text = studentGrid[6, index].Value.ToString();
+                panel4.Visible = true;
+            }
+        }
+
+        private void button12_Click(object sender, EventArgs e)
+        {
+            try
+            {
+                if (label19.Text == "Добавление")
+                {
+                    if ((groupStudentBox.Text == "") || (textBox9.Text == "") || (textBox10.Text == "") || (textBox8.Text == "") || (textBox7.Text == "") || (maskedTextBox4.Text == ""))
+                    {
+                        MessageBox.Show("Заполните все поля");
+                    }
+                    else
+                    {
+                        int j = studentGrid.RowCount - 2;
+                        string query = $"INSERT INTO Student(idGroup, lastName, firstName, sureName, address, phoneNum) VALUES ({groupStudentBox.SelectedValue},'{textBox9.Text}','{textBox10.Text}','{textBox8.Text}', '{textBox7.Text}', '{maskedTextBox4.Text}')";
+                        OleDbCommand command = new OleDbCommand(query, _connection);
+                        command.ExecuteNonQuery();
+                        panel4.Visible = false;
+                    }
+                }
+                else
+                if (label19.Text == "Изменение")
+                {
+                    int index = studentGrid.CurrentRow.Index;
+                    if (index == studentGrid.RowCount - 1)
+                    {
+                        MessageBox.Show("Выберите заполненную сторку");
+                    }
+                    else
+                    {
+                        int id = Convert.ToInt32(studentGrid[0, index].Value.ToString());
+                        string query = $"UPDATE Student SET idGroup = {groupStudentBox.SelectedValue}, lastName = '{textBox9.Text}', firstName = '{textBox10.Text}', sureName = '{textBox8.Text}', address = '{textBox7.Text}', phoneNum = '{maskedTextBox4.Text}' WHERE Student.id = {id}";
+                        OleDbCommand command = new OleDbCommand(query, _connection);
+                        command.ExecuteNonQuery();
+                        StudentShow();
+                        panel4.Visible = false;
+                    }
+                }
+                StudentShow();
+            }
+            catch (Exception a)
+            {
+                MessageBox.Show(a.Message);
+            }
+}
+
+        private void button1_Click(object sender, EventArgs e)
+        {
+            var res = MessageBox.Show("Вы действительно хотите удалить запись?", "Удаление", MessageBoxButtons.YesNo, MessageBoxIcon.Exclamation);
+            if (res == DialogResult.Yes)
+            {
+                panel4.Visible = false;
+                int index = studentGrid.CurrentRow.Index;
+                if (index == studentGrid.RowCount - 1)
+                {
+                    MessageBox.Show("Выберите заполненную сторку");
+                }
+                else
+                {
+                    int id = Convert.ToInt32(studentGrid[0, index].Value.ToString());
+                    string query = $"DELETE FROM Student WHERE Student.id = {id}";
+                    OleDbCommand command = new OleDbCommand(query, _connection);
+                    command.ExecuteNonQuery();
+                }
+                StudentShow();
+            }
+        }
     }
 }
